requirements.txt
fastapi
uvicorn
pymongo
pydantic
jwt
passlib
python-dotenv
